layout playlist song selection component
