+++ src/components/Playlist/Search/Search.tsx
@@ -1,0 +1,16 @@
+// Libraries
+import React from 'react'
+import { withRouter } from 'react-router-dom'
+
+// Styles
+import './Search.css';
+
+const Search = withRouter((props) => {  
+  return (
+    <div className="Search">
+      <p>Search</p>
+    </div>
+  )
+});
+
+export default Search
